requirements.txt
# Core Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Async HTTP Client
aiohttp>=3.9.0
httpx>=0.26.0

# Configuration
python-dotenv>=1.0.0
pyyaml>=6.0

# Logging
structlog>=24.1.0
python-json-logger>=2.0.7

# Retry & Utils
tenacity>=8.2.3

# Monitoring (Optional)
prometheus-client>=0.19.0

# Development
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.12.0
ruff>=0.1.9
